+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change `ScheduleAppointmentSheet` to use Mobius view effects
 - Add Clubhouse overview description in the Pivotal documentation
 - Migrate v1 `Country` usages to v2 `Country`
+- Change `NewMedicalHistoryScreen` to use Mobius view effects
 - [In Progress: 30 Aug 2021] Record call results instead of updating the same appointment record
 
 ### Changes